Exception hendling in employee service comleted. Need to do more validation in pojo.

+++ njProjectServer/src/main/java/com/njProjectServer/service/EmployeeService.java
@@ -2,6 +2,7 @@ package com.njProjectServer.service;
 
 import com.njProjectServer.exception.ResourceNotFoundException;
 import com.njProjectServer.model.*;
+import com.njProjectServer.model.dto.InsertAssistantDto;
 import com.njProjectServer.model.dto.InsertEmployeeDto;
 import com.njProjectServer.model.dto.InsertTeacherDto;
 import com.njProjectServer.repository.*;
@@ -63,25 +64,80 @@ public class EmployeeService {
         return teachersRepository.save(tech);
     }
 
-    public Assistant insertAssistant(Assistant assistant) {
-        return assistantRepository.save(assistant);
+    public Assistant insertAssistant(InsertAssistantDto assistant) {
+        Optional<Title> t = titleRepository.findById(assistant.getTitleID());
+
+        if(t.isEmpty())
+            throw new ResourceNotFoundException("Title with given id not found!");
+
+        Assistant a = new Assistant(assistant.getName(), assistant.getSurname(), assistant.getEmploymentDate(), t.get());
+
+        return assistantRepository.save(a);
     }
 
     public void delete(int id){
+        Optional<Employee> e = employeeRepository.findById(id);
+        if(e.isEmpty())
+            throw new ResourceNotFoundException("Employee with given id not found");
         employeeRepository.deleteById(id);
     }
 
 
-    public Employee update(Employee employee) {
-        return employeeRepository.save(employee);
+    public Employee update(InsertEmployeeDto employee, int id) {
+        Optional<Employee> emp = employeeRepository.findById(id);
+        if (emp.isEmpty()){
+            throw new ResourceNotFoundException("Employee with given id not found");
+        }
+        emp.get().setEmploymentDate(employee.getEmploymentDate());
+        emp.get().setName(employee.getName());
+        emp.get().setSurname(employee.getSurname());
+
+        return employeeRepository.save(emp.get());
     }
 
-    public Assistant updateAssistant(Assistant assistant) {
-        return assistantRepository.save(assistant);
+    public Assistant updateAssistant(int id, InsertAssistantDto assistant) {
+        Optional<Assistant> a = assistantRepository.findById(id);
+        if (a.isEmpty()){
+            throw new ResourceNotFoundException("Assistant with given id not found");
+        }
+        a.get().setEmploymentDate(assistant.getEmploymentDate());
+        a.get().setName(assistant.getName());
+        a.get().setSurname(assistant.getSurname());
+
+        Optional<Title> title = titleRepository.findById(assistant.getTitleID());
+
+        if(title.isEmpty())
+            throw new ResourceNotFoundException("Title with given id not found!");
+
+        a.get().setTitle(title.get());
+
+        return assistantRepository.save(a.get());
     }
 
-    public Teacher updateTeacher(Teacher teacher) {
-        return teachersRepository.save(teacher);
+    public Teacher updateTeacher(int id,InsertTeacherDto teacher) {
+        Optional<Teacher> t = teachersRepository.findById(id);
+        if (t.isEmpty()){
+            throw new ResourceNotFoundException("Teacher with given id not found");
+        }
+        t.get().setEmploymentDate(teacher.getEmploymentDate());
+        t.get().setName(teacher.getName());
+        t.get().setSurname(teacher.getSurname());
+
+        Optional<Title> title = titleRepository.findById(teacher.getTitleID());
+
+        if(title.isEmpty())
+            throw new ResourceNotFoundException("Title with given id not found!");
+
+        t.get().setTitle(title.get());
+
+        Optional<Rank> rank = rankRepository.findById(teacher.getRankID());
+
+        if(rank.isEmpty())
+            throw new ResourceNotFoundException("Rank with given id not found!");
+
+        t.get().setRank(rank.get());
+
+        return teachersRepository.save(t.get());
     }
 
     public ResponseEntity<Employee> findById(int id) {
